Hack: Don't reset the current file list when replaced
I want to be able to start multiple gThumb sessions with different
command lines without the previous ones instantly clearing. Granted, the
concept of a stored, global catalog starts to break down at that point,
but as long as we can pretend that our catalog hasn't been replaced for
a little while, I'm happy.