image viewer task: Fix infinite recursion bug
If original_image_task_completed_cb is called with a non-null error, it
attempts to propogate the error. But it called gth_task_completed on the
child task that just failed instead of the parent task, signaling itself
and entering an infinite recursion. Fix this by singnaling completion of
the correct task.