image viewer: do not mess with the current file
If the current file shows up in the browser after loading it, this
causes the current file position to suddenly change. If it's saved to a
different location, the current file position gets reset and the viewer
jumps to the beginning if the user attempts to advance to the next file.

Leaving the current file alone means the viewer's position will stay
where it is, which is less surprising.

Also, mark the file as unmodified after a successful save-as to prevent
asking the user whether to save it, since they already did.